% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puckR-package.R
\docType{package}
\name{puckR-package}
\alias{puckR}
\alias{puckR-package}
\title{puckR: Collect and Clean Hockey Stats}
\description{
Various functions to scrape hockey play-by-play data from the
\url{https://www.nhl.com/}. It also contains functions to scrape data from \url{https://www.hockey-reference.com/},
including standings, player stats, and jersey number history.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zbeg/puckR}
  \item Report bugs at \url{https://github.com/zbeg/puckR/issues}
}

}
\author{
\strong{Maintainer}: Daniel Morse \email{zaiemb@gmail.com}

}
\keyword{internal}
